Add support for SystemTap profile script output
Add a command-line tool and reader for dealing with output of
SystemTap profiling scripts.

To capture data using SystemTap, you can use:

$ stap -e "global ubt; \
           probe timer.profile { ubt[execname(), sprint_ubacktrace()] += 1 }; \
           probe timer.s(30) { exit() }" \
       -o stap.out
$ cmd/stackcollapse-stap < stap.out > collapsed.out

If stap warns about missing unwind data for a module, and stap
suggests adding '-d /lib/libquux.so', which you know to be a shared
library used by the 'foo' binary, add the following to the stap
command above:

       -d /path/to/foo $(ldd /path/to/foo | awk 'NF==4 { print "-d", $3 }')

You can add other stap expressions as keys to the ubt[] array, as
long as sprint_ubacktrace() remains the last key in the list.
Values of the other expressions will be added as prefixes to the
frames in the collapsed stacks, for example:

ubt[execname(), sprint_ubacktrace()]

and running some binary 'foo' which calls functions 'bar' and 'baz',
will output foo's stack frames as:

foo`bar,foo`baz

in the collapsed stacks output.